overlay: Set a wake lock while running Waydroid
Opportunistic autosuspend tends to block waydroid-sensors from processing events
from sensorfw, leading to crashes and unrecoverable restarts of Waydroid, or worse,
lockups in repowerd's sensorfw SocketReader, leading to screens not turning back on
again. To counter that, create a Waydroid LXC config_base with hooks at start and stop
of the container to set and unset a wake lock while running Waydroid. This ensures
the system stays awake while Waydroid is up and running, allowing sensor events
to be processed in a timely manner.